refactor: expr editor
