+++ frontend/src/components/CustomApproval/Settings/components/common/index.ts
@@ -1,4 +1,3 @@
-import ExprEditor from "./ExprEditor";
 import RequiredStar from "./RequiredStar.vue";
 import RiskFilter from "./RiskFilter";
 import SpinnerButton from "./SpinnerButton.vue";
@@ -7,10 +6,8 @@ import SpinnerSwitch from "./SpinnerSwitch.vue";
 
 export * from "./utils";
 export * from "./RiskFilter";
-export * from "./ExprEditor";
 export {
   RequiredStar,
-  ExprEditor,
   SpinnerButton,
   SpinnerSwitch,
   SpinnerSelect,
