+++ frontend/src/components/ExprEditor/ExprEditor.vue
@@ -5,9 +5,9 @@
 </template>
 
 <script lang="ts" setup>
+import type { SelectOption } from "naive-ui";
 import { toRef } from "vue";
-import type { ConditionGroupExpr } from "@/plugins/cel";
-import { Risk_Source } from "@/types/proto/v1/risk_service";
+import type { ConditionGroupExpr, Factor } from "@/plugins/cel";
 import ConditionGroup from "./ConditionGroup.vue";
 import { provideExprEditorContext } from "./context";
 
@@ -15,12 +15,12 @@ const props = withDefaults(
   defineProps<{
     expr: ConditionGroupExpr;
     allowAdmin?: boolean;
-    allowHighLevelFactors?: boolean;
-    riskSource?: Risk_Source;
+    factorList: Factor[];
+    factorSupportDropdown: Factor[];
+    factorOptionsMap: Map<Factor, SelectOption[]>;
   }>(),
   {
     allowAdmin: false,
-    allowHighLevelFactors: false,
     riskSource: undefined,
   }
 );
@@ -31,8 +31,9 @@ defineEmits<{
 
 provideExprEditorContext({
   allowAdmin: toRef(props, "allowAdmin"),
-  allowHighLevelFactors: toRef(props, "allowHighLevelFactors"),
-  riskSource: toRef(props, "riskSource"),
+  factorList: toRef(props, "factorList"),
+  factorSupportDropdown: toRef(props, "factorSupportDropdown"),
+  factorOptionsMap: toRef(props, "factorOptionsMap"),
 });
 </script>
